requirements.txt
langchain>=0.3.0
requests>=2.28.0
python-dotenv>=1.0.0
wikipedia-api>=0.5.8
duckduckgo-search>=5.3.1
tqdm>=4.65.0
colorama>=0.4.6
flask>=3.0.0
flask-cors>=4.0.0
